packages/backend/pyproject.toml

[project]
name = "freedom-label-backend"
dynamic = [ "version" ]
license = "Apache-2.0"
authors = [
    { name = "Nicola Abis" },
]
readme = "README.md"

[tool.poetry]
name = "freedom-label-backend"
version = "0.1.0"
description = ""
authors = ["Gemini <gemini@google.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
prometheus-fastapi-instrumentator = "^6.1.0"
fpdf2 = "^2.8.3"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
mypy = "^1.10.0"
types-fpdf2 = "^2.8.3.20250809"
trio = "^0.30.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
logger-objects = ["pkg.module.logger"]
ignore = ["FBT001", "FBT002", "FIX002", "D203", "D213"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
"_initial_attempts/**/*.py" = ["ALL"]

[tool.mypy]
strict = true
exclude_gitignore = true
exclude = [
  '.tests\/*\.py$',
  '_initial_attempts'
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
